feature (extended translucency): ScatteringVolume material
Adding a new material model to the feature: ScatteringVolume
This material is modeled after a ball of smoke, which scattering light exponentially
through it.

It is unique because its the first material model that "fade out" the mesh,
instead of strengthening its edges.

This material was originally designed for `Ghost`, or `SetActorAlpha`.
But it does not work out of the box yet:
1. The EffectShader of a regular ghost is way too strong, making it hard to see difference in the mesh's rendering.
2. We disabled this effect in SKIN shader to avoid rendering issue with makeup and face details.